Bugfix: queuing rays which require more regions than the cache can hold
Fix an issue which enqueuing regions when the GPU cache is too small to hold all the regions for a single ray set.

This is a bit of a messy fix, but the existing logic really didn't hold up when the cache was too small to fit all the regions for selected rays.

The fix will flush *all* in flight invocations (previously just the previous buffer) and try again. There is some messy management around re-using the same rays buffer (previously copying between GPU buffers).